send functions for polynomials from utilities to linear_algebra
