+++ src/ore_algebra/analytic/linear_algebra.py
@@ -33,7 +33,7 @@ from sage.functions.other import binomial
 
 from .complex_optimistic_field import ComplexOptimisticField
 from .accuracy import PrecisionError
-from .utilities import roots, XGCD
+from .utilities import derivatives
 
 
 
@@ -58,7 +58,7 @@ def customized_accuracy(x):
 
     EXAMPLES::
 
-        sage: from ore_algebra.analytic.utilities import customized_accuracy
+        sage: from ore_algebra.analytic.linear_algebra import customized_accuracy
         sage: a = ComplexBallField().one()
         sage: a.accuracy(), customized_accuracy(a)
         (9223372036854775807, 9223372036854775807)
@@ -83,6 +83,237 @@ def customized_accuracy(x):
         return max(0, (-log(x.rad(), 2)).floor())
 
     return x.accuracy()
+
+
+
+
+################################################################################
+### Polynomials avec ComplexOptimisticField ####################################
+################################################################################
+
+def _clean(pol):
+
+    l = list(pol)
+    while len(l)>0 and l[-1].contains_zero(): l.pop()
+    cpol = pol.parent()(l)
+
+    return cpol
+
+def GCD(a, b):
+
+    r"""
+    Return a *non-rigorous* gcd of the polynomials "a" and "b".
+
+    Note: this function is designed for BallField as base ring.
+
+    Some words about the correction of this function:
+    Let a· and b· be fixed. If a and b are precise enough, GCD(a, b) contains
+    the gcd of a· and b·.
+
+    INPUT:
+
+     -- "a" -- polynomial
+     -- "b" -- polynomial
+
+    OUTPUT:
+
+     -- "a" -- polynomial
+
+
+    EXAMPLES::
+
+        sage: from ore_algebra.analytic.linear_algebra import GCD
+        sage: P.<x> = CBF[]; a = CBF(pi)
+        sage: p, q = (x-1)*(x-2)**2, (x-2)*(x-3)**2
+        sage: p, q = p(x*a), q(x*a)
+        sage: d = GCD(p, q); d(x/a).monic()
+        ([1.0000000000 +/- 1.34e-12])*x + [-2.00000000000 +/- 1.94e-12]
+
+    """
+
+    a, b = _clean(a), _clean(b)
+    if a==0: return b
+    if b==0: return a
+    if a.degree() < b.degree(): return GCD(b, a)
+
+    while b != 0:
+        a, b = b, a.quo_rem(b)[1]
+        b = _clean(b)
+
+    return a
+
+
+
+def XGCD(a, b):
+
+    r"""
+    Return a *non-rigorous* monic gcd of the polynomials "a" and "b" and the
+    coefficients in the Bezout identity.
+
+    Note: this function is designed for BallField as base ring.
+
+    INPUT:
+
+     -- "a" -- polynomial
+     -- "b" -- polynomial
+
+    OUTPUT:
+
+     -- "d" -- polynomial
+     -- "u" -- polynomial
+     -- "v" -- polynomial
+
+
+    EXAMPLES::
+
+        sage: from ore_algebra.analytic.linear_algebra import XGCD, _clean
+        sage: P.<x> = CBF[]; a = CBF(pi)
+        sage: p, q = (x-1)*(x-2)**2, (x-2)*(x-3)**2
+        sage: p, q = p(x*a), q(x*a)
+        sage: d, u, v = XGCD(p, q); d, 2/a
+        (([1.0000000000 +/- 1.33e-12])*x + [-0.63661977237 +/- 3.04e-12],
+         [0.636619772367581 +/- 4.28e-16])
+        sage: _clean(u*p + v*q)
+        ([1.000000000 +/- 3.39e-11])*x + [-0.63661977237 +/- 6.95e-12]
+
+    """
+
+    P = a.parent()
+
+    a, b = _clean(a), _clean(b)
+    if a==0: return b, P.zero(), P.one()
+    if b==0: return a, P.one(), P.zero()
+    if a.degree() < b.degree():
+        d, v, u = XGCD(b, a)
+        return d, u, v
+
+    r0, u0, v0, r1, u1, v1 = a, P.one(), P.zero(), b, P.zero(), P.one()
+    while r1!=0:
+        r0, (q, r1) = r1, r0.quo_rem(r1)
+        u0, v0, u1, v1 = u1, v1, u0 - q*u1, v0 - q*v1
+        r1 = _clean(r1)
+
+    lc = r0.leading_coefficient()
+    d, u, v = r0.monic(), _clean(u0/lc), _clean(v0/lc)
+
+    return d, u, v
+
+
+
+def squarefree_part(pol):
+
+    r"""
+    Return a *non-rigorous* squarefree part of the polynomial "pol".
+
+    Note: this function is designed for BallField as base ring.
+
+    Some words about the correction of this function:
+    Let pol· be fixed. If pol is precise enough, squarefree_part(pol) contains
+    the squarefree part of pol·.
+
+
+    INPUT:
+
+     -- "pol" -- polynomial
+
+
+    OUTPUT:
+
+     -- "sfp" -- polynomial
+
+
+    EXAMPLES::
+
+        sage: from ore_algebra.analytic.linear_algebra import squarefree_part
+        sage: P.<x> = CBF[]; a = CBF(pi)
+        sage: p = (x-1)*(x-2)**2
+        sage: p = p(x*a).monic(); p
+        ([1.0000000000000 +/- 1.57e-15])*x^3 + ([-1.59154943091895 +/- 5.18e-15])*x^2 + ([0.81056946913870 +/- 2.96e-15])*x + [-0.129006137732798 +/- 1.12e-16]
+        sage: sfp = squarefree_part(p); sfp
+        ([-44.4132198049 +/- 2.39e-11])*x^2 + ([42.4115008235 +/- 7.69e-11])*x + [-9.0000000000 +/- 3.38e-11]
+        sage: sfp.roots(multiplicities=False)
+        [[0.3183098862 +/- 2.31e-11] + [+/- 6.86e-12]*I,
+         [0.6366197724 +/- 4.20e-11] + [+/- 9.55e-12]*I]
+        sage: [1/a, 2/a]
+        [[0.318309886183791 +/- 4.43e-16], [0.636619772367581 +/- 4.28e-16]]
+
+    """
+
+    d = GCD(pol, pol.derivative())
+    sfp = _clean(pol.quo_rem(d)[0])
+
+    if sfp==0:
+        raise PrecisionError("Cannot compute the squarefree part of this polynomial.")
+
+    return sfp
+
+
+
+def roots(pol, *, multiplicities=False):
+
+    r"""
+    Return the roots of the polynomial "pol".
+
+    Note: this function is designed for CBF or COF as base ring.
+
+    Some words about the correction of this algorithm:
+
+
+    INPUT:
+
+     -- "mat"            -- n×n matrix
+     -- "multiplicities" -- boolean
+
+
+    OUTPUT:
+
+     -- "s" -- list of complex numbers
+
+    If 'multiplicities=True' is specified, "s" is a list of couples (r, m) with
+    r a complex number and m a positive integer.
+
+
+    EXAMPLES::
+
+        sage: from ore_algebra.analytic.linear_algebra import roots
+        sage: P.<x> = CBF[]; a = CBF(pi)
+        sage: p = (x-1)*(x-2)**2; p = p(x*a).monic()
+        sage: roots(p, multiplicities=True)
+        [([0.3183098862 +/- 2.31e-11] + [+/- 6.86e-12]*I, 1),
+         ([0.6366197724 +/- 4.20e-11] + [+/- 9.55e-12]*I, 2)]
+        sage: p.roots(multiplicities=False,proof=False)
+        <ipython-input-7-b3a160b9a5ea>:1: UserWarning: roots may have been lost
+          p.roots(multiplicities=False,proof=False)
+        [[0.3183098861838 +/- 2.69e-14] + [+/- 1.75e-14]*I,
+         [0.636620 +/- 8.62e-7] + [+/- 6.48e-7]*I,
+         [0.636620 +/- 8.62e-7] + [+/- 6.48e-7]*I]
+
+    """
+
+    K, n = pol.base_ring(), pol.degree()
+    if isinstance(K, ComplexOptimisticField):
+        pol = pol.change_ring(K._ball_field)
+
+    try:
+        res = squarefree_part(pol).roots(multiplicities=False)
+        res = [K(r) for r in res]
+    except ValueError:
+        raise PrecisionError("Cannot compute the roots of this polynomial.") from None
+
+    if not multiplicities: return res
+
+    for j, ev in enumerate(res):
+        m = 1
+        evaluations = [p(ev) for p in derivatives(pol, n)]
+        while evaluations[m].contains_zero():
+            m = m + 1
+        res[j] = (ev, m)
+
+    if sum(m for _, m in res)<n:
+        raise PrecisionError("Cannot compute multiplicities.")
+
+    return res
+
